+++ gemrb/core/GameData.h
@@ -39,8 +39,6 @@
 #include "SrcMgr.h"
 #include "TableMgr.h"
 
-#include "Scriptable/Actor.h"
-
 #include <map>
 #include <vector>
 
@@ -53,6 +51,7 @@ class ScriptedAnimation;
 class Sprite2D;
 class Store;
 class VEFObject;
+enum class Difficulty;
 
 struct IWDIDSEntry {
 	ieDword value;
