+++ gemrb/core/Interface.cpp
@@ -24,6 +24,7 @@
 #include "exports.h"
 #include "globals.h"
 #include "ie_cursors.h"
+#include "ie_stats.h"
 #include "strrefs.h"
 
 #include "ActorMgr.h"
