+++ gemrb/core/SaveGameIterator.cpp
@@ -20,6 +20,7 @@
 
 #include "SaveGameIterator.h"
 
+#include "ie_stats.h"
 #include "strrefs.h"
 
 #include "DisplayMessage.h"
