+++ gemrb/core/Scriptable/Scriptable.h
@@ -22,34 +22,22 @@
 #define SCRIPTABLE_H
 
 #include "exports.h"
-#include "ie_cursors.h"
 #include "ie_types.h"
-#include "strrefs.h"
 
-#include "CharAnimations.h"
 #include "OverHeadText.h"
 #include "PathFinder.h"
 
 #include <list>
 #include <map>
-#include <memory>
 
 namespace GemRB {
 
 class Action;
 class Actor;
-class Container;
-class Door;
 class GameScript;
-class Highlightable;
-class InfoPoint;
 class Map;
-class Movable;
 class Object;
-struct PathNode;
 class Projectile;
-class Scriptable;
-class Selectable;
 class Spell;
 
 #define MAX_GROUND_ICON_DRAWN 3
