+++ gemrb/plugins/FXOpcodes/FXOpcodes.cpp
@@ -18,7 +18,9 @@
  *
  */
 
+#include "defsounds.h"
 #include "ie_feats.h" //cannot avoid declaring these
+#include "ie_stats.h"
 #include "opcode_params.h"
 #include "overlays.h"
 #include "strrefs.h"
