+++ gemrb/plugins/IWDOpcodes/IWDOpcodes.cpp
@@ -19,6 +19,7 @@
  */
 
 #include "ie_feats.h" //cannot avoid declaring these
+#include "ie_stats.h"
 #include "opcode_params.h"
 #include "overlays.h"
