+++ gemrb/plugins/TLKImporter/TLKImporter.cpp
@@ -20,6 +20,8 @@
 
 #include "TLKImporter.h"
 
+#include "ie_stats.h"
+
 #include "Calendar.h"
 #include "DialogHandler.h"
 #include "Game.h"
